This is a major reorg of HILTI's internals, changing the code structure quite a bit and generaly cleaning things up. The diff is big and thus suppressed.
There are also some news. HILTI now supports

- timers

- hash maps and sets, with automatic item expiration.

- printing out ASTs as parseable HILTI source (was only partially
  working before).

[nodiff]

+++ hilti/libhilti/memory.c
@@ -8,32 +8,32 @@
 
 #include "hilti.h"
 
-void* hlt_gc_malloc_atomic(size_t n)
+void* hlt_gc_malloc_atomic(uint64_t n)
 {
     return malloc(n);
 }
 
-void* hlt_gc_malloc_non_atomic(size_t n)
+void* hlt_gc_malloc_non_atomic(uint64_t n)
 {
     return malloc(n);
 }
 
-void* hlt_gc_calloc_atomic(size_t count, size_t n)
+void* hlt_gc_calloc_atomic(uint64_t count, uint64_t n)
 {
     return calloc(count, n);
 }
 
-void* hlt_gc_calloc_non_atomic(size_t count, size_t n)
+void* hlt_gc_calloc_non_atomic(uint64_t count, uint64_t n)
 {
     return calloc(count, n);
 }
 
-void* hlt_gc_realloc_atomic(void* ptr, size_t n)
+void* hlt_gc_realloc_atomic(void* ptr, uint64_t n)
 {
     return realloc(ptr, n);
 }
 
-void* hlt_gc_realloc_non_atomic(void* ptr, size_t n)
+void* hlt_gc_realloc_non_atomic(void* ptr, uint64_t n)
 {
     return realloc(ptr, n);
 }
